Refactoring union fields for better compiler errors
